Moved H5VolumeSelectionDlg to stackFileSelectionWidget.py
To make the dialog available in both the stackfile selection and the
dataSelection, this class was moved.
